[Fix] 캘린더 오늘 날짜 밑줄 튀는 이슈 수정, 캘린더뷰 UI 디테일 수정 (#85)
