+++ SobokSobok/SobokSobok.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -33,7 +33,7 @@
         "repositoryURL": "https://github.com/Taehyeon-Kim/EasyKit.git",
         "state": {
           "branch": "master",
-          "revision": "9766f933ee840c840104ebe4a78c18183db7a640",
+          "revision": "5861ae869104f7ac5443dfa5bc91178d08ded9ac",
           "version": null
         }
       },
