+++ SobokSobok/SobokSobok/Presentation/Home/Calendar/CalendarViewController.swift
@@ -7,13 +7,10 @@
 
 import UIKit
 
+import EasyKit
 import FSCalendar
 
-final class CalendarViewController: BaseViewController, PageComponentProtocol {
-    var pageTitle: String {
-        "수현"
-    }
-
+final class CalendarViewController: BaseViewController {
     // MARK: - UI
     @IBOutlet weak var scrollView: UIScrollView!
     
@@ -27,62 +24,49 @@ final class CalendarViewController: BaseViewController, PageComponentProtocol {
 
     @IBOutlet weak var collectionView: UICollectionView!
     @IBOutlet weak var collectionViewHeight: NSLayoutConstraint!
+
+    // MARK: - Properties
+    /// Data
+    var doingDates = [String]()
+    let doneDates = [String]()
+    var selectedDate: String = Date().toString(of: .day) {
+        didSet {
+            dateLabel.text = selectedDate
+        }
+    }
     
-    var tabType: TabBarItem = .share
-    let someDates = ["2022-01-09", "2022-01-22", "2022-01-30"]
-    let allDates = ["2022-01-12", "2022-01-15", "2022-01-17"]
+    /// Item
+    var scheduleItems: [String] = []
+    var pillItems: [String] = []
+    
+    let gregorian = Calendar(identifier: .gregorian)
+    private var calendarExpandedState: Bool = false {
+        didSet {
+            calendarExpandedState ?
+            calendar.setScope(.month, animated: true) :
+            calendar.setScope(.week, animated: true)
+            scopeLabel.text = calendarExpandedState ? "월" : "주"
+        }
+    }
     var editMode: Bool = false {
         didSet {
             collectionView.reloadData()
         }
     }
     
-    // MARK: - Properties
-    
-    private var calendarAnimatedState: Bool = false {
-        didSet {
-            calendarAnimatedState ?
-            calendar.setScope(.month, animated: true) :
-            calendar.setScope(.week, animated: true)
-            scopeLabel.text = calendarAnimatedState ? "월" : "주"
-        }
-    }
-    
-    fileprivate let gregorian = Calendar(identifier: .gregorian)
-    fileprivate let formatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "yyyy-MM-dd"
-        return formatter
-    }()
-    fileprivate let kformatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "ko_KR")
-        formatter.dateFormat = "M월 d일 EEEE"
-        return formatter
-    }()
-    
-    var selectedDate: String = "" {
-        didSet {
-            dateLabel.text = selectedDate
-        }
-    }
-    
-    // 데이터라고 가정
-    fileprivate var pillNames: [String] = ["홍삼", "비타민C", "루테인", "혈압약", "알로에", "당뇨약"] {
-        didSet {
-            collectionView.reloadData()
-        }
-    }
+    var tabType: TabBarItem = .share
+    var tabName: String = "수현"
     
     // MARK: - Life Cycles
     
     override func viewDidLoad() {
         super.viewDidLoad()
+        
+        registerXibs()          // 등록 먼저
+        setDelegation()
         setCalendar()
         setCalendarStyle()
         setCollectionView()
-        scrollView.delegate = self
-        usernameStackView.isHidden = tabType == .home
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -95,26 +79,32 @@ final class CalendarViewController: BaseViewController, PageComponentProtocol {
     }
 
     override func style() {
-        super.style() // 오버라이딩 위해서는 부모 뷰 상속 필요
+        super.style()
+        
+        usernameStackView.isHidden = tabType == .home
         usernameLabel.setTypoStyle(typoStyle: .header1)
         dateLabel.setTypoStyle(typoStyle: .title2)
         scopeLabel.setTypoStyle(typoStyle: .body7)
-        selectedDate = kformatter.string(from: Date())
     }
     
-    // MARK: - Functions
+    @IBAction func scopeButtonTapped(_ sender: Any) {
+        calendarExpandedState.toggle()
+    }
+}
+
+// MARK: - Functions
+
+extension CalendarViewController {
+    // MARK: - Set
     
     private func setCalendar() {
-        calendar.delegate = self
-        calendar.dataSource = self
+        calendar.locale = Locale(identifier: "ko_KR")
         calendar.headerHeight = 0
+        calendar.firstWeekday = 2
         calendar.setScope(.week, animated: false)
-        calendar.register(CalendarDayCell.self, forCellReuseIdentifier: CalendarDayCell.reuseIdentifier)
     }
     
     private func setCalendarStyle() {
-        calendar.locale = Locale(identifier: "ko_KR")
-        calendar.firstWeekday = 2
         calendar.appearance.weekdayFont = TypoStyle.body7.font
         calendar.appearance.weekdayTextColor = Color.gray600
         calendar.appearance.titleFont = UIFont.font(.pretendardReular, ofSize: 16)
@@ -125,204 +115,86 @@ final class CalendarViewController: BaseViewController, PageComponentProtocol {
     }
     
     private func setCollectionView() {
-        collectionView.delegate = self
-        collectionView.dataSource = self
         collectionView.backgroundColor = Color.gray150
-        collectionView.register(MedicineCollectionViewCell.self)
-        collectionView.register(TimeHeaderView.nib,
-                                forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
-                                withReuseIdentifier: TimeHeaderView.reuseIdentifier)
-        
+
+        // FlowLayout
         let flowLayout: UICollectionViewFlowLayout = UICollectionViewFlowLayout()
+        flowLayout.minimumLineSpacing = 8
         let screenSize = UIScreen.main.bounds.width
         let sideMargin = 40 / 375 * screenSize
-        flowLayout.estimatedItemSize = CGSize(width: collectionView.frame.width - sideMargin, height: 140)
+        flowLayout.estimatedItemSize = CGSize(width: screenSize - sideMargin, height: 500)
         collectionView.collectionViewLayout = flowLayout
         
         self.view.layoutIfNeeded()
     }
     
-    @IBAction func scopeButtonTapped(_ sender: Any) {
-        calendarAnimatedState.toggle()
+    private func registerXibs() {
+        calendar.register(
+            CalendarDayCell.self,
+            forCellReuseIdentifier: CalendarDayCell.reuseIdentifier
+        )
+        collectionView.register(MedicineCollectionViewCell.self)
+        collectionView.register(
+            TimeHeaderView.nib,
+            forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
+            withReuseIdentifier: TimeHeaderView.reuseIdentifier
+        )
     }
     
-    private func showActionSheet() {
+    /// 컬렉션뷰 높이 조정 함수
+    public func setCollectionViewHeight() {
+        collectionViewHeight.constant = collectionView.contentSize.height
+    }
+    
+    private func setDelegation() {
+        calendar.delegate = self
+        calendar.dataSource = self
+        collectionView.delegate = self
+        collectionView.dataSource = self
+        scrollView.delegate = self
+    }
+    
+    // MARK: - Helpers
+
+    public func showActionSheet() {
         let actionSheet = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
         let editAction = UIAlertAction(title: "약 수정", style: .default) { _ in }
         let stopAction = UIAlertAction(title: "복약 중단", style: .default) { _ in }
         let deleteAction = UIAlertAction(title: "약 삭제", style: .default) { _ in }
         let cancelAction = UIAlertAction(title: "취소", style: .cancel)
-        actionSheet.addAction(editAction)
-        actionSheet.addAction(stopAction)
-        actionSheet.addAction(deleteAction)
-        actionSheet.addAction(cancelAction)
+        
+        actionSheet.addAction(editAction, stopAction, deleteAction, cancelAction)
         present(actionSheet, animated: true, completion: nil)
     }
-}
-
-// MARK: - FSCalendar
-
-extension CalendarViewController: FSCalendarDelegate {
-    func calendar(_ calendar: FSCalendar, boundingRectWillChange bounds: CGRect, animated: Bool) {
-        calendarHeight.constant = bounds.height
-        self.view.layoutIfNeeded()
-    }
-}
-
-extension CalendarViewController: FSCalendarDataSource {
-    func calendar(_ calendar: FSCalendar, cellFor date: Date, at position: FSCalendarMonthPosition) -> FSCalendarCell {
-        guard let cell = calendar.dequeueReusableCell(
-            withIdentifier: CalendarDayCell.reuseIdentifier,
-            for: date,
-            at: position
-        ) as? CalendarDayCell else { return FSCalendarCell() }
-        configure(cell: cell, for: date, at: position)
-        return cell
-    }
     
-    func calendar(_ calendar: FSCalendar, willDisplay cell: FSCalendarCell, for date: Date, at monthPosition: FSCalendarMonthPosition) {
-
-    }
-    
-    private func configure(cell: FSCalendarCell, for date: Date, at position: FSCalendarMonthPosition) {
-        guard let cell = cell as? CalendarDayCell else { return }
-        
-        var filledType = FilledType.none
-        var selectedType = SelectedType.not
-        var width: CGFloat = 32
-        var yPosition: CGFloat = 2
-        let formattedDate = formatter.string(from: date)
-        
-        if self.gregorian.isDateInToday(date) {
-            filledType = .today
-        } else if someDates.contains(formattedDate) {
-            filledType = .some
-        } else if allDates.contains(formattedDate) {
-            filledType = .all
-        } else {
-            filledType = .none
+    public func showStickerBottomSheet() {
+        let stickerBottomSheet = StickerBottomSheet.instanceFromNib()
+        stickerBottomSheet.modalPresentationStyle = .overCurrentContext
+        stickerBottomSheet.modalTransitionStyle = .crossDissolve
+        self.tabBarController?.present(stickerBottomSheet, animated: false) {
+            stickerBottomSheet.showSheetWithAnimation()
         }
-        
-        if calendar.selectedDates.contains(date) && !gregorian.isDateInToday(date) {
-            width = 44
-            yPosition = -4
-            selectedType = .single
-        }
-        
-        cell.width = width
-        cell.yPosition = yPosition
-        cell.selectedType = selectedType
-        cell.filledType = filledType
-    }
-}
-
-extension CalendarViewController: FSCalendarDelegateAppearance {
-    func calendar(_ calendar: FSCalendar, didSelect date: Date, at monthPosition: FSCalendarMonthPosition) {
-        pillNames = pillNames.reversed()
-        selectedDate = kformatter.string(from: date)
-        
-        calendar.select(date)
-        self.configureVisibleCells()
-    }
-    
-    private func configureVisibleCells() {
-        calendar.visibleCells().forEach { (cell) in
-            let date = calendar.date(for: cell)
-            let position = calendar.monthPosition(for: cell)
-            self.configure(cell: cell, for: date!, at: position)
-        }
-    }
-}
-
-// MARK: - CollectionView
-
-extension CalendarViewController: UICollectionViewDelegate, UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        3
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        4
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(for: indexPath, cellType: MedicineCollectionViewCell.self)
-        cell.contentView.backgroundColor = Color.white
-        cell.contentView.makeRounded(radius: 12)
-        cell.pillName.text = pillNames[indexPath.row % 6]
-        
-        // 테스트 위한 코드
-        if indexPath.row % 2 == 0 {
-            cell.stickerStackView.isHidden = true
-            cell.stickerCountLabel.isHidden = true
-        }
-        
-        cell.editButton.isHidden = !editMode
-        cell.checkButton.isHidden = editMode
-        
-        cell.stickerClosure = { [weak self] in
-            guard let self = self else { return }
-                
-            let stickerBottomSheet = StickerBottomSheet.instanceFromNib()
-            stickerBottomSheet.modalPresentationStyle = .overCurrentContext
-            stickerBottomSheet.modalTransitionStyle = .crossDissolve
-            self.tabBarController?.present(stickerBottomSheet, animated: false) {
-                stickerBottomSheet.showSheetWithAnimation()
-            }
-        }
-        
-        cell.editClosure = {
-            self.showActionSheet()
-        }
-        
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        switch kind {
-        case UICollectionView.elementKindSectionHeader:
-            guard let headerView = collectionView.dequeueReusableSupplementaryView(
-                ofKind: kind,
-                withReuseIdentifier: TimeHeaderView.reuseIdentifier,
-                for: indexPath
-            ) as? TimeHeaderView else { return UICollectionReusableView() }
-            if indexPath.section == 0 {
-                headerView.editButtonStackView.isHidden = false
-            } else {
-                headerView.editButtonStackView.isHidden = true
-            }
-            headerView.editModeClosure = {
-                self.editMode.toggle()
-            }
-            return headerView
-        default:
-            assert(false, "헤더 뷰 찾을 수 없음")
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-        let width: CGFloat = collectionView.frame.width
-        let height: CGFloat = 77
-        return CGSize(width: width, height: height)
     }
 }
 
 extension CalendarViewController: UICollectionViewDelegateFlowLayout {
-    func setCollectionViewHeight() {
-        collectionViewHeight.constant = collectionView.contentSize.height
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        8
-    }
-    
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        UIEdgeInsets(top: 0, left: 20, bottom: 32, right: 20)
+        if section == 2 {
+            return UIEdgeInsets(top: 0, left: 10, bottom: 100, right: 10)
+        } else {
+            return UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 10)
+        }
     }
 }
 
 extension CalendarViewController: UIScrollViewDelegate {
     func scrollViewDidScroll(_ scrollView: UIScrollView) {
         scrollView.backgroundColor  = scrollView.contentOffset.y > UIScreen.main.bounds.height / 2 ? Color.gray150 : Color.white
+    }
+}
+
+extension CalendarViewController: PageComponentProtocol {
+    var pageTitle: String {
+        return tabName
     }
 }
