+++ SobokSobok/SobokSobok/Presentation/Home/HomeViewController.xib
@@ -35,7 +35,7 @@
                     <state key="normal" image="icMenu48"/>
                 </button>
                 <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="hT0-FU-Tjs">
-                    <rect key="frame" x="0.0" y="124.66666666666669" width="375" height="653.33333333333326"/>
+                    <rect key="frame" x="0.0" y="146.66666666666669" width="375" height="631.33333333333326"/>
                     <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                 </view>
             </subviews>
@@ -48,7 +48,7 @@
                 <constraint firstItem="hT0-FU-Tjs" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="Tsl-VM-TDH"/>
                 <constraint firstItem="vAo-UM-pJe" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="28" id="ZRj-qt-SOB"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="hT0-FU-Tjs" secondAttribute="trailing" id="aQM-ba-u7T"/>
-                <constraint firstItem="hT0-FU-Tjs" firstAttribute="top" secondItem="nhP-yl-zeR" secondAttribute="bottom" id="tKO-tO-cXV"/>
+                <constraint firstItem="hT0-FU-Tjs" firstAttribute="top" secondItem="nhP-yl-zeR" secondAttribute="bottom" constant="22" id="tKO-tO-cXV"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="bottom" secondItem="hT0-FU-Tjs" secondAttribute="bottom" id="zKB-8V-8Wk"/>
             </constraints>
             <point key="canvasLocation" x="139" y="147"/>
